pyproject.toml
# SPDX-FileCopyrightText: 2025 Stefano Miccoli <stefano.miccoli@polimi.it>
#
# SPDX-License-Identifier: MIT

[project]
name = "suanpan-abaqus"
dynamic = ["version"]
description = "Python interface to Abaqus .fil files"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
readme = "README.md"
authors = [
    { name = "Stefano Miccoli", email = "stefano.miccoli@polimi.it" }
]
requires-python = ">=3.10"
dependencies = ["numpy>=1.21.3"]
license = "MIT"
license-files = ["LICENCES/MIT.txt"]

[project.scripts]
filinfo = "suanpan.cli.filinfo:main"

[project.optional-dependencies]
filinfo = [ "PyYAML >= 6.0.3" ]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

#
# hatch
#
[tool.hatch.version]
source = "vcs"
[tool.hatch.version.raw-options]
version_scheme = "release-branch-semver"
local_scheme = "node-and-date"
[tool.hatch.build.hooks.vcs]
version-file = "src/suanpan/_version.py"
[tool.hatch.build.targets.wheel]
packages = ["src/suanpan"]

#
# ruff
#
[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = ["A", "EM", "I", "NPY", "S", "SIM", "TRY", "UP"]
ignore = ["S101"]
